backend/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
psycopg2-binary
google-generativeai
python-dotenv
